Add JWT authentication and Swagger configuration
This commit introduces JWT authentication to the project, including the addition of a new `JwtConfig.cs` file for configuring authentication parameters. The `SwaggerConfigExtensions.cs` file is added to support Swagger documentation with JWT authentication. The `TokenService.cs` constructor is updated for dependency injection of `ITokenService`. Several related package references are added to `OrderProcess.Infrastructure.csproj`, and unused using directives are removed for code cleanup.